reducing work load
git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@3156 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ code/base/dso-common-jdk15/tests.unit/com/tc/net/groups/TribesGroupManagerTest.java
@@ -176,7 +176,7 @@ public class TribesGroupManagerTest extends TCTestCase {
   private void checkMessagesOrdering(final TribesGroupManager mgr1, final TribesGroupManager mgr2)
       throws GroupException {
 
-    final Integer upbound = new Integer(100);
+    final Integer upbound = new Integer(50);
     final MyListener myl1 = new MyListener();
     final MyListener myl2 = new MyListener();
     mgr1.registerForMessages(TestMessage.class, myl1);
